requirements.txt
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-MCP3008==1.0.2
Adafruit-PureIO==1.1.9
anyio==3.5.0
asgiref==3.5.0
click==8.1.0
colorzero==2.0
fastapi==0.75.0
greenlet==1.1.2
guizero==1.3.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
RPi.GPIO==0.7.1
sniffio==1.2.0
spidev==3.5
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
